Move test discovery to appmap-agent-init
Do test discovery in appmap-agent-init instead of appmap-agent-status.
appmap-agent-status now reads test commands from appmap.yml.
