+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Test discovery now happens in `appmap-agent-init` instead of `appmap-agent-status`.
+- Use a virtual env to test `appmap-agent-init`'s test discovery.
+
 ## [1.2.0.dev1] - 2021-07-16
 ### Added
 - The `appmap-agent-init` and `appmap-agent-status` internal commands are now available
